Polygon is rendering on the map. Still need to size the map appropriately and display the users location.
