+++ .gitignore
@@ -1,3 +1,5 @@
-/node_modules/*
+/node_modules
 .env
-/dist/*
+/public/*.js
+/build/*.js
+/.bak
