--- bs-config.json
@@ -1,8 +1,0 @@
-{
-  "server": {
-    "baseDir": "./public",
-    "routes": {
-      "/dist": "dist"
-    }
-  }
-}
