+++ package.json
@@ -6,19 +6,18 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "tsc && npm run minify",
-    "minify": "terser src/*.js --output dist",
-    "start": "lite-server"
+    "minify": "npx uglifyjs build/index.js --compress --mangle --output public/index.min.js",
+    "start-dev": "http-server -c-1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@googlemaps/js-api-loader": "^1.16.2",
-    "@types/googlemaps": "^3.43.3"
+    "http-server": "^14.1.1",
+    "uglify-js": "^3.17.4"
   },
   "devDependencies": {
-    "@types/google.maps": "^3.54.10",
-    "lite-server": "^2.6.1",
-    "terser": "^5.26.0"
+    "@types/google.maps": "^3.54.10"
   }
 }
