+++ public/style.css
@@ -6,11 +6,10 @@ body {
   padding: 0;
   background-color: #fff;
   font-family: 'Roboto', sans-serif;
-  font-size: 16px;
+  font-size: 18px;
   font-weight: 400;
   line-height: 1.5;
   color: #212529;
-  text-align: left;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
@@ -28,16 +27,10 @@ h2 {
   font-size: 2rem;
 }
 
-/* required-location-map-outer */
-#required-location-map-outer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+#map-outer {
   min-height: 300px;
   border: 1px solid #ccc;
 }
-#required-location-map {
-  width: 100%;
+#map {
   height: 100%;
 }
