--- src/util.ts
@@ -1,27 +1,0 @@
-// the max distance in meters that the user can be from the geofence
-const maxDistance = 100;
-
-// Define geofence coordinates
-const geofenceCoordinates = [
-  {
-    lng: -149.9041557,
-    lat: 61.2190324
-  },
-  {
-    lng: -149.8588371,
-    lat: 61.2187845
-  },
-  {
-    lng: -149.8542023,
-    lat: 61.1877758
-  },
-  {
-    lng: -149.9093056,
-    lat: 61.1876103
-  },
-  {
-    lng: -149.9038124,
-    lat: 61.2191977
-  }
-  // Add more coordinates as needed for your geofence zones
-];
